Rename weatherNowSlice to weatherSlice
